heathen and pysword_repo
lots added to pysword_repo, boot strapping function, and a handful of other functions related to that.

Working notes and such in HR config and reader mode and most everything else, but setting up new first run check and boot strapping and havent tested so it may(most likely) not work atm.